requirements.txt
# Web & Environment
fastapi
uvicorn
python-dotenv
python-multipart
pydantic
httpx

# AI Orchestration (2026 Unified Google GenAI)
llama-index-core
llama-index-llms-google-genai
llama-index-embeddings-google-genai
llama-index-vector-stores-qdrant
google-genai

# Infrastructure & Tools
qdrant-client
supabase
pymupdf